Plugins: categorize linux.check_modules as a malware plugin
